Update the add new comment features
